Initial commit
This is going to be the main code-repository for the chaosinventory
project. This commit brings a basic django-project structure with an
hello world template. It is structured to be distributed and installed
"as is" without creating an additional django project. It was heavily
inspired by the @pretix project[pretix] and I often peaked at their
approach to the setup.py, app-structure and so on.

This is, just as stated in the setup.py, in a pre-alpha state. Major
TODOs include:

* Reading preferences from a config file
  * Pythons `configparser` seems a suitable fit
  * Pretix could be used as an example, especially for the logical
    structure of it, they're doing an awesome job of it in my opinion!
* Creating the models from our documentation
* Build the application itself, this is just a starting point

[pretix]: https://pretix.eu