Add initial API ViewSets, Serializers and Routes
This is not complete yet, not all Models are available trough the API,
there has been no though given on how we will need to interact with the
API for which features. For now, all of them are simple ReadWrite
ViewSets[^viewsets] which permit the listing, creation, (retrival,
updating, deletion) of single objects.

Currently, there is no documentation.

When visiting `/api/` you will be presented with a browsable
representation of the API which is rely nice during development, but
curl request will get valid json, and by appending `.json` to a path you
can see it in your browser to. Further explanation can be found here:
https://www.django-rest-framework.org/topics/browsable-api/

[^viewsets]: https://www.django-rest-framework.org/api-guide/viewsets/