Add lint.sh script
Currently checking isort and flake8 and exit 1 of one or both of them
fail.

Ina ddition to the output of these linters, a small status is printed
like this:

  isort:	[pass|fail]
  flake8:	[pass|fail]

If isort fails, this message will be printed out:

  For an interactive auto fixer run
    isort --interactive .
  from /path/to/chaosinventory/src/

This script can be linked as a git hook but can also be run manually.
Documentation is still todo and I will do that later on.

Close #37