Human readable error if invalid configfile is provided (#56)
If no configuration was provided (detected if there are no sections)
this error gets logged:

    WARNING:root:No custom configuration file was provided! Continuing with example values.

When the configparser throws an error (e.g. syntax is invlid) the error
will be formatted nicely and print some additional information. For
example, the errors doc-string is prinetd out as a help.

This may look like this:

    CRITICAL:root:DuplicateOptionError occured while parsing the configuration at chaosinventory.cfg:3:
    While reading from 'chaosinventory.cfg' [line  3]: option 'secret' in section 'django' already exists
    Note: Raised by strict parsers when an option is repeated in an input source.

Close #16