Allow certain values to be blank
Tags should not be required to create an object as not all objects will have a need for a tag.
Setting the "null"-property is not required for the tag fields as the representation in the db doesn't have fields which could be nulled

The parent field has to be blankable to allow for creation of one or multiple "root"-objects.