Lint the Linters
This commit introduced so called CheckFlavours. Checks can now use
placeholders like `[[path]]` that will be substituted by values that can
be defined centrally in the check flavours. If the flavours name is not
default, it will be appended in brackets to the printed name of the
check.

Running a specific check (e.g. flake8) will trigger all flavours of this
check.

Close #57