Ensure Tag / Overlay can not parent themself
We need to set a custom `save()` function since Django lets the
application handle validation (as in calling `full_clean()`) by default.
This also affects the validation of e.g. `blank=True`, which is not
enforced outside of the AdminUI (and possibly the API)

Resources about djangos validation process:

- https://stackoverflow.com/a/4441740/11249686
- https://stackoverflow.com/a/18876223/11249686
- https://docs.djangoproject.com/en/3.1/ref/models/instances/#django.db.models.Model.clean
- https://gist.github.com/glarrain/5448253

Close #62