Bootstrap REST API (#31)
* Add djangorestframework requirements

Ref #10

* Split views file and add django_extensions

- By splitting the views into multiple files we can keep them more
  organized and clean when the API with a load of view will come in.
- django_extensions is a neat helper during development, when this
  feature is final it could be removed but for now I'd like to keep it
  as a helper.

* Add initial API ViewSets, Serializers and Routes

This is not complete yet, not all Models are available trough the API,
there has been no though given on how we will need to interact with the
API for which features. For now, all of them are simple ReadWrite
ViewSets[^viewsets] which permit the listing, creation, (retrival,
updating, deletion) of single objects.

Currently, there is no documentation.

When visiting `/api/` you will be presented with a browsable
representation of the API which is rely nice during development, but
curl request will get valid json, and by appending `.json` to a path you
can see it in your browser to. Further explanation can be found here:
https://www.django-rest-framework.org/topics/browsable-api/

[^viewsets]: https://www.django-rest-framework.org/api-guide/viewsets/

* Add some documentation for the API

This should give a rough run-over the tecinal structure of the API.

* Apply suggestions from code review

Mostly some typos  or small inconstent things for now.

Co-Authored-By: Leo Maroni <git@em0lar.de>

* Update documentation arround available HTTP methods

Co-Authored-By: Leo Maroni <git@em0lar.de>

* Reflect changes from #24 in InventoryIds

* Add ItemData_set to ItemSerializer

Small refracture of the *Data_set serializers since they are quite
simmelar.

* Document how to export a openapi schema

Also added the appropriate requirements

* Fix a lot of stuff

- Removed recursive serialisation (see doc-blocks and docs) since
  creation of Items would not be possible without implementin custom
  create methods;
- Added ItemInventoryIdSerializer and ProductInventoryIdSerializer with
  their corresponding ViewSets and routes.

* Fix autodoc builds

By including the dev-requirements the django rest framework and all
other depedencies are present. This is easier than manually keeping the
dependencies in sync.

* Add Overlay{,Item} to API

* Add {Product,Item}Data Serializers and endpoints

* Add misssing endpoints to docs

* Make overlayitem_set read_only

Implements changes requests in code review.

Co-Authored-By: Leo Maroni <git@em0lar.de>

Co-authored-by: Leo Maroni <git@em0lar.de>