Do not overwrite save() but add a custom validator
In our matrix chat, @clerie mentioned that overwriting the save method
is not that great due to various reasons (performance, reducing
flexibility in the code, ...).

This I reworked the Serializers to a custom validate function. However
we can't, as much as I would have liked to, use validators (e.g. the
clean function) of the Model itself.

For a deeper explanation of why, see the doc-block of the
TagSerializer.validate().

I am not sattisfied with the current state where we effectively have
validation logic in the validate function instead of the models, but I
dont't know how we could improve on this...

Co-Authored-By: clerie <git@clerie.de>