Make API resolve objects recursively
This will show all related objects in e.G. list requests. If a model is
referencing itself, it will be resolved one layer deep. However, if it
is not referencing itself all of the ForeignKeys will be resolved and
their ForeignKeys will to. I tried to avoid that e.g. a Item Objects
will appear n-layers deep again and cause infinite recursion.

For every serializer with foreign keys (so almost all of them) I created
multiple "generations". There are three groups of serializers and their
generations.

1. Generation serializers: They do not have any foreign keys, thus can
   be used as is.
2. Generation serializers: They have foreign keys, but are only required
   in a writable state (PKs instead of resolved objects) and a fully
   resolved state (All FKs are mapped to the propper serializer). The
   writable generation is prefixed with `Base` (`BasicItemSerializer`)
   and the fully resolved generation has no prefix (`ItemSerializer`).
3. Generation serializers: The writable (first) generation is prefixed
   with `Writable` (`WritableLocationDataSerializer`). The second
   generation with `Base` (`BasicLocationDataSerializer`) and the final,
   as before, has no prefix (`LocationDataSerializer`). This generation
   is used in only a couple of cases. For example, in the
   final `LocationDataSerializer` we want to both resolve the location
   and the `DataType`. But in the `LocationSerializer` (=> the Location
   list) we already have the location and only need the type resolved.
   Therefore, in the `LocationSerializer` the `LocationData` Objects are
   resolved to the `BasicLocationDataSerializer`. For writing to the
   Database however, we can not have any resolved foreign keys[^0] thus
   the `WritableLocationDataSerializer` without and resolved
   relationships is used.

Usually generations inherit from the previous version since the fields
do not change, but some of them get resolved.

Instead of, as initially thought, overwriting the create class, the
different serializer generations are handled in the ViewSets. Using the
`AbstractModelViewSet` (may be renamed in the future), the create and
update methods are overwritten to use `self.writable_serializer_class`.
In the future, this wrapper could be expanded to provide easy
customization for example to use different serializers in the list and
detail endpoints (for less recursion in the list, and thus save data,
but extend a lot more in the single objects view).

One issue is, that the generated schema[^1] is not compatible with the
resolved  lists and detail views (the spec only describes the ids being
returned while objects are given). Implementing all of that to show up
in the schema is a big TODO, but required for proper API documentation.

The API can be tested using postman[^2], with the schema import all
endpoints and their parameters will automatically show up. The list /
detail views will throw erros however, due to the before mentioned
issues with the schema.

Close #67

[^0]: To be fair, it may be possible but I did some testing and was not
successful to write anything fully working, let alone having easy to
understand code.
[^1]: `§ python manage.py generateschema --format openapi > schema.yml`
[^2]: <https://www.postman.com/>