Ensure objects pointing on same class can not parent themself (#65)
* Ensure Tag / Overlay can not parent themself

We need to set a custom `save()` function since Django lets the
application handle validation (as in calling `full_clean()`) by default.
This also affects the validation of e.g. `blank=True`, which is not
enforced outside of the AdminUI (and possibly the API)

Resources about djangos validation process:

- https://stackoverflow.com/a/4441740/11249686
- https://stackoverflow.com/a/18876223/11249686
- https://docs.djangoproject.com/en/3.1/ref/models/instances/#django.db.models.Model.clean
- https://gist.github.com/glarrain/5448253

Close #62

* Do not overwrite save() but add a custom validator

In our matrix chat, @clerie mentioned that overwriting the save method
is not that great due to various reasons (performance, reducing
flexibility in the code, ...).

This I reworked the Serializers to a custom validate function. However
we can't, as much as I would have liked to, use validators (e.g. the
clean function) of the Model itself.

For a deeper explanation of why, see the doc-block of the
TagSerializer.validate().

I am not sattisfied with the current state where we effectively have
validation logic in the validate function instead of the models, but I
dont't know how we could improve on this...

Co-Authored-By: clerie <git@clerie.de>

* Sketching a circular dependency check

* Renaming function to make consitent

* Implement parent checking in all other models too

* Reverting unused method for debugging

Co-authored-by: clerie <git@clerie.de>