Add OIDC authentication method
In this commit OIDC authentication method is initialized.
Currently supported are:
* Authorization Flow
* Authorization Flow with PKCE
* Endpoints:
  * Authorize
  * Token with issuing Access, Refresh and ID Tokens

Also a simple login template was added, as it is needed for OIDC
authorize endpoint.

This commit also introduces breaking internal and external changes:
* External: All authentication endpoints moved from
`/api/authentication/...` to `/auth/`. The reason for that change is
that the OIDC endpoints aren't API endpoints.
* Internal: The folder structure of `/src/chaosinventory/authentication`
has changed to fit to new requirements.

Co-Authored-By: Moritz 'e1mo' Fromm <git@e1mo.de>