Implement unit tests (#77)
* Add some initial model tests

The current tests in tests/base/test_models.py mostly check the
constraints that objects can not have a parental (or comparable)
relationship with them self.

The tests can be run using `python manage.py test tests` from the src
directory.

Ref #76

* Run tests in GitHub Actions

* Fix no return in parental checks in base.models

* Test the token endpoint and authentication

* Test renewed token workin

Start using a standartized API Endpoint (GET /api/authentication/token/)
to test, if the current token is valid.

* Fix token endpoint

Usually, API endpoints end on /, however this one does not. This leads
to issues and test fails.

* Add unit tests to pre-commit-hook

* Fix copy and paste in gh actions

* Add small and sample api tag test

* Fix failing Tag API creation test

* Add more token related tests

- test_token_bad_password
- test_token_delete
  - Currently fails due to missing / in the endpoint => fix in next
    commit
- test_token_new
  - Limited, see #79
- test_token_expiry
  - Taking 1.1 seconds (due to a short wait before the token expires)
  - Paralel tests could reduce that time but currently they do not work

* Add missing / to token detail endpoint

* Add simple coverage testing

Currently, this is just some additional output in the action log. In the
future we _could_ push it to something like codecov.io[^0] for better
visualization / feedback on PRs.

[^0]: https://docs.codecov.io/docs

* Add some basic docs on the unit tests.

* Remove testdummy from coverage exclude

* Add note to test_user_token_in_db

* Apply requested changes from review

Co-Authored-By: Leo Maroni <git@em0lar.de>

* Deduplicate api request with a small mixin

* Mock time instead of using sleep

Implement changes requested by @margau

Co-authored-by: Leo Maroni <git@em0lar.de>